requirements.txt
alipay==0.7.4
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
Django==2.1.5
django-core==1.4.1
django-haystack==2.8.1
django-redis==4.10.0
django-redis-sessions==0.5.6
django-tinymce==2.6.0
docopt==0.6.2
fdfs-client-py==1.2.6
get==2018.11.19
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==4.2.2.post1
mutagen==1.42.0
mysql==0.0.2
mysqlclient==1.3.14
Pillow==5.4.1
pipreqs==0.4.9
pkg-resources==0.0.0
post==2018.11.20
public==2018.11.20
pycryptodomex==3.7.2
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.7
query-string==2018.11.20
redis==2.10.6
redis-py-cluster==1.3.6
request==2018.11.20
requests==2.21.0
six==1.12.0
urllib3==1.24.1
uWSGI==2.0.18
vine==1.1.4
Whoosh==2.7.4
yarg==0.1.9
